requirements.txt
pyrogram
TgCrypto
py-tgcalls
python-dotenv
youtube_dl
youtube_search_python
aiofiles
aiohttp
requests
asyncio
youtube_search
search_engine_parser
wget
yt_dlp
ffmpeg
Pillow
ujson
telethon
cryptg
heroku3
dnspython
cryptg
requests
